requirements.txt
snowflake-connector-python==3.6.0
pandas==2.1.4
python-dotenv==1.0.0
streamlit==1.32.0
